pull out issueMany into file for re-use
