+++ test/RToken.test.ts
@@ -35,6 +35,7 @@ import {
   SLOW,
 } from './fixtures'
 import { cartesianProduct } from './utils/cases'
+import { issueMany } from './utils/issue'
 
 const createFixtureLoader = waffle.createFixtureLoader
 
@@ -1158,44 +1159,6 @@ describe(`RTokenP${IMPLEMENTATION} contract`, () => {
     return basketHandler.status()
   }
 
-  // Issue `amount` RToken to `user`.
-  //
-  // Because hardhat_mine actually takes O(n) time to mine n
-  // blocks, we do this more cleverly than just one big issuance...
-  // This presumes that user has already granted allowances of basket tokens!
-  async function issue_many(toIssue: BigNumber, user: SignerWithAddress): Promise<void> {
-    const ISS_BLOCKS = bn(1e3) // How many blocks to wait between issuances; tweak to tune performance
-    const MIN_ISSUANCE_RATE = fp(10000) // {rtoken / block}
-
-    let supply = await rToken.totalSupply()
-    let issued = bn(0)
-    const initBalance = await rToken.balanceOf(user.address)
-    const issuanceRate = await rToken.issuanceRate()
-
-    while (issued.lt(toIssue)) {
-      // Find currIssue, the amount to issue this round
-      const yetToIssue = toIssue.sub(issued)
-      const baseAmt = MIN_ISSUANCE_RATE.mul(ISS_BLOCKS)
-      const succAmt = supply.mul(issuanceRate).mul(ISS_BLOCKS)
-      const maxAmt = baseAmt.gt(succAmt) ? baseAmt : succAmt
-      const currIssue = maxAmt.lt(yetToIssue) ? maxAmt : yetToIssue
-
-      // Issue currIssue to user, and wait ISS_BLOCKS
-
-      await rToken.connect(user).issue(currIssue)
-
-      await advanceBlocks(ISS_BLOCKS.add(1))
-
-      await rToken.vest(user.address, await rToken.endIdForVest(user.address))
-      issued = issued.add(currIssue)
-      supply = supply.add(currIssue)
-    }
-
-    // assert that this worked
-    expect(await rToken.balanceOf(user.address)).to.equal(initBalance.add(toIssue))
-    expect(await rToken.totalSupply()).to.equal(supply)
-  }
-
   describe(`Extreme Values (${SLOW ? 'slow mode' : 'fast mode'})`, () => {
     async function runScenario([
       toIssue,
@@ -1263,13 +1226,13 @@ describe(`RTokenP${IMPLEMENTATION} contract`, () => {
       // ==== Issue the "initial" rtoken supply to owner
 
       expect(await rToken.balanceOf(owner.address)).to.equal(bn(0))
-      await issue_many(toIssue0, owner)
+      await issueMany(rToken, toIssue0, owner)
       expect(await rToken.balanceOf(owner.address)).to.equal(toIssue0)
 
       // ==== Issue the toIssue supply to addr1
 
       expect(await rToken.balanceOf(addr1.address)).to.equal(0)
-      await issue_many(toIssue, addr1)
+      await issueMany(rToken, toIssue, addr1)
       expect(await rToken.balanceOf(addr1.address)).to.equal(toIssue)
 
       // ==== Send enough rTokens to addr2 that it can redeem the amount `toRedeem`
